Donald's Contribution
chatcube_chatroom.html is working if you're only going to talk to
yourselves. The next step would be to link it to the server. Ryan I was
thinking you could do it. Once that is done I can add some more
functionality. Also there are 3 font sizes to choose from small, medium
and large.